Moved more things around, lint passes.
