+++ experiments/exp1994_32b_sft.py
@@ -9,13 +9,12 @@ and uses the mixture frmo 1880
 import dataclasses
 
 from fray.cluster import ResourceConfig
-from marin.defaults import default_sft
+from marin.defaults import SimpleSFTConfig, default_sft
 from marin.execution.executor import executor_main
 
 from experiments.evals.evals import default_sft_eval
 from experiments.exp1880_sft_baseline import mixture_config
 from experiments.marin_models import marin_tokenizer
-from experiments.simple_sft_config import SimpleSFTConfig
 from experiments.tootsie.exp1529_32b_mantis_cooldown import (
     qwen3_32b_remat,
     tootsie_32b_cooldown_mantis,
