+++ experiments/exp606_sft.py
@@ -9,11 +9,11 @@ This module provides the `tulu3_llama_data_old` dataset configuration and
 """
 
 from fray.cluster import ResourceConfig
+from marin.defaults import SimpleSFTConfig
 from marin.processing.tokenize import lm_data_config
 
 from experiments.llama import llama3_instruct_chat_format, llama3_instruct_tokenizer
 from experiments.posttrain.instruction_datasets import get_instruction_dataset
-from experiments.simple_sft_config import SimpleSFTConfig
 from experiments.tokenization import default_tokenize
 
 # Get instruction dataset
