+++ experiments/exp808_sft_mixture.py
@@ -12,7 +12,7 @@ accordingly.
 
 from fray.cluster import ResourceConfig
 from levanter.data.text import DEFAULT_LM_DATA_SHUFFLE, ChatLmDatasetFormat
-from marin.defaults import default_sft
+from marin.defaults import SimpleSFTConfig, default_sft
 from marin.execution.executor import executor_main
 from marin.execution.types import ExecutorStep
 from marin.processing.tokenize import lm_mixture_data_config
@@ -20,7 +20,6 @@ from marin.processing.tokenize import lm_mixture_data_config
 from experiments.llama import llama_8b
 from experiments.marin_models import marin_tokenizer
 from experiments.posttrain.instruction_datasets import get_instruction_dataset
-from experiments.simple_sft_config import SimpleSFTConfig
 from experiments.tokenization import default_tokenize
