--- experiments/simple_sft_config.py
@@ -1,167 +1,0 @@
-# Copyright The Marin Authors
-# SPDX-License-Identifier: Apache-2.0
-
-from dataclasses import dataclass
-
-from fray.cluster import ResourceConfig
-from levanter.schedule import IntSchedule
-
-
-def compute_per_device_parallelism(
-    global_batch_size: int,
-    microbatch_size: int,
-    resources: ResourceConfig,
-) -> int:
-    """Compute per_device_parallelism for gradient accumulation.
-
-    Args:
-        global_batch_size: The effective batch size after gradient accumulation.
-        microbatch_size: The batch size that fits in memory (local batch size).
-        resources: The ResourceConfig specifying TPU/GPU resources.
-
-    Returns:
-        per_device_parallelism: Number of examples each device processes per forward/backward pass.
-
-    Example:
-        For v5p-8 (4 chips), global_batch_size=128, microbatch_size=8:
-        - per_device_parallelism = 8 / 4 = 2
-        - gradient_accumulation = 128 / 8 = 16 steps
-    """
-    num_devices = resources.chip_count()
-
-    if microbatch_size % num_devices != 0:
-        raise ValueError(f"microbatch_size ({microbatch_size}) must be divisible by " f"num_devices ({num_devices})")
-
-    if global_batch_size % microbatch_size != 0:
-        raise ValueError(
-            f"global_batch_size ({global_batch_size}) must be divisible by " f"microbatch_size ({microbatch_size})"
-        )
-
-    per_device_parallelism = microbatch_size // num_devices
-    grad_accum_steps = global_batch_size // microbatch_size
-
-    print(
-        f"Gradient accumulation config: "
-        f"global_batch={global_batch_size}, microbatch={microbatch_size}, "
-        f"num_devices={num_devices}, per_device_parallelism={per_device_parallelism}, "
-        f"grad_accum_steps={grad_accum_steps}"
-    )
-
-    return per_device_parallelism
-
-
-@dataclass(frozen=True)
-class SimpleSFTConfig:
-    """
-    A simplified configuration for Supervised Fine-Tuning (SFT) that works for both
-    single dataset and mixture training approaches.
-    """
-
-    # Hardware configuration
-    resources: ResourceConfig
-
-    # Core training parameters
-    train_batch_size: int | IntSchedule = 128
-    """
-    The batch size for training. If an IntSchedule is provided, the batch size will be
-    varied according to the schedule.
-    """
-    num_train_steps: int = 10000
-    """Number of training steps."""
-
-    learning_rate: float = 5e-6
-    """Learning rate for the optimizer."""
-
-    # Model configuration
-    tokenizer: str | None = None
-    """Tokenizer to use for training."""
-
-    initialize_from_hf: str | None = None
-    """HF model name or path to initialize from (e.g., 'meta-llama/Llama-3.1-8B').
-    Mutually exclusive with initialize_from_checkpoint_path."""
-
-    initialize_from_checkpoint_path: str | None = None
-    """Path to a levanter checkpoint to initialize from.
-    Mutually exclusive with initialize_from_hf."""
-
-    max_seq_len: int = 4096
-    """Maximum sequence length for training."""
-
-    # Optimizer parameters
-    weight_decay: float = 0.0
-    """Weight decay for the optimizer."""
-
-    beta1: float | None = None
-    """AdamW optimizer beta1."""
-
-    beta2: float | None = None
-    """AdamW optimizer beta2."""
-
-    warmup: float = 0.03
-    """Fraction of training steps to use for learning rate warmup."""
-
-    decay: float = 0.0
-    """Fraction of training steps to use for learning rate decay."""
-
-    lr_schedule: str = "linear"
-    """Learning rate schedule to use: 'linear', 'cosine', etc."""
-
-    min_lr_ratio: float = 0.0
-    """Minimum learning rate as a ratio of the base learning rate."""
-
-    max_grad_norm: float | None = None
-    """Maximum gradient norm for gradient clipping."""
-
-    # Checkpointing and evaluation
-    steps_per_eval: int = 1000
-    """How often to run validation losses."""
-
-    steps_per_checkpoint: int | None = None
-    """How often to keep a permanent checkpoint. None (default) keeps only the final
-    checkpoint; rolling temporary checkpoints are still written for resumption."""
-
-    steps_per_hf_export: int = 500
-    """How often to save HuggingFace checkpoints."""
-
-    hf_generation_eos_token_ids: list[int] | None = None
-    """EOS token IDs to write to generation_config.json. None means no generation config.
-    For chat models, include the turn-boundary token (e.g. [128001, 128009])."""
-
-    # Mixture-specific parameters
-    mixture_block_size: int = 2048
-    """Block size for dataset mixing (only used with mixture training)."""
-
-    stop_strategy: str = "restart"
-    """
-    Strategy for handling dataset completion (only used with mixture training).
-    Options: 'restart' or 'exit'.
-    """
-
-    # Other parameters
-    seed: int = 0
-    """Random seed for training."""
-
-    node_count: int = 1
-    """Number of TPU slices for training."""
-
-    int8: bool = False
-    """Int8 (quantized) training in Levanter."""
-
-    pad_tokenizer_to_match_model: bool = False
-    """If True, pad the tokenizer's vocab to match the model's vocab size by adding dummy tokens.
-    Useful when the model checkpoint has a larger vocab than the tokenizer (e.g., Qwen models
-    pad their vocab to be divisible by 4 for TPU efficiency)."""
-
-    z_loss_weight: float = 0.0
-
-    per_device_parallelism: int = -1
-    """How many examples to process in parallel on each device. -1 (default) means
-    train_batch_size/num_devices (no gradient accumulation). Set to a positive value
-    to enable gradient accumulation. For example, with 8 devices, batch_size=32, and
-    per_device_parallelism=1, you get gradient accumulation of 4."""
-
-    reinit_tokens: list[str] | bool = False
-    """
-    if set, will reinitialize the embeddings for the given tokens. If True, will reinitialize the default tokens
-    for llama3's tokenizer
-    """
