+++ experiments/test_dpo_generation_config.py
@@ -13,14 +13,13 @@ Expected content: {"eos_token_id": [128001, 128009], "bos_token_id": 128000}
 
 from fray.cluster import ResourceConfig
 from levanter.data.text import PreferenceChatLmDatasetFormat
-from marin.defaults import default_dpo
+from marin.defaults import SimpleDPOConfig, default_dpo
 from marin.execution.executor import executor_main
 from marin.processing.tokenize import lm_data_config
 
 from experiments.llama import LLAMA3_CHAT_STOP_TOKEN_IDS, llama_8b
 from experiments.marin_models import marin_tokenizer
 from experiments.posttrain.preference_datasets import get_preference_dataset
-from experiments.simple_dpo_config import SimpleDPOConfig
 from experiments.tokenization import default_tokenize
 
 DATASET_NAME = "HuggingFaceH4/ultrafeedback_binarized"
