+++ tests/test_training.py
@@ -15,7 +15,7 @@ from levanter.data.text import DatasetComponent, PreferenceChatLmDatasetFormat, 
 from levanter.main import train_lm
 from levanter.main.train_dpo import TrainDpoConfig
 from levanter.trainer import TrainerConfig
-from marin.defaults import default_dpo
+from marin.defaults import SimpleDPOConfig, default_dpo
 from marin.processing.tokenize import tokenized_cache_stats_path
 from marin.training.training import (
     TrainDpoOnPodConfig,
@@ -28,7 +28,6 @@ from marin.training.training import (
 )
 
 from experiments.llama import llama_8b
-from experiments.simple_dpo_config import SimpleDPOConfig
 from experiments.tokenization import default_tokenize
